+++ mods/sbz_progression/init.lua
@@ -158,6 +158,7 @@ local achievement_table = {
     ["sbz_chem:engraver_off"] = "Engraver",
     ["sbz_multiblocks:blast_furnace_controller"] = "Blast Furnace",
     ["sbz_instatube:instant_tube"] = "Instatubes",
+    ["sbz_power:teleport_battery"] = "Teleport Battery"
 }
 
 minetest.register_on_craft(function(itemstack, player, old_craft_grid, craft_inv)